requirements.txt
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
numpy==1.18.4
pandas==1.0.3
prometheus-client==0.7.1
pycodestyle==2.6.0
python-dateutil==2.8.1
python-gitlab==2.2.0
pytz==2020.1
requests==2.23.0
six==1.15.0
urllib3==1.25.9
